+++ src/main/scala/players/AI2.scala
@@ -1,4 +1,7 @@
-package game
+package main.scala.players
+
+import main.scala.elements.{GridOfAttack, GridOfShips, Ship}
+
 import scala.util.Random
 
 /** An AI2 for a battleship game
